+++ cropclassification/taskrunner.py
@@ -80,7 +80,9 @@ def run_tasks(tasksdir: Path, config_overrules: list[str] = []):
         action = task_config["task"].get("action")
         default_basedir = None
         if "paths" in task_config:
-            default_basedir = Path(task_config["paths"].get("marker_basedir"))
+            marker_basedir = task_config["paths"].get("marker_basedir")
+            if marker_basedir is not None:
+                default_basedir = Path(marker_basedir)
         if default_basedir is None:
             default_basedir = task_path.parent.parent
 
@@ -109,6 +111,14 @@ def run_tasks(tasksdir: Path, config_overrules: list[str] = []):
             from cropclassification import calc_cropclass
 
             calc_cropclass.run_cropclass(
+                config_paths=config_paths,
+                default_basedir=default_basedir,
+                config_overrules=config_overrules,
+            )
+        elif action == "calc_cover":
+            from cropclassification import calc_cover
+
+            calc_cover.run_cover(
                 config_paths=config_paths,
                 default_basedir=default_basedir,
                 config_overrules=config_overrules,